delete: 나와의 경쟁 페이지 삭제
