+++ apps/mac/src/renderer/src/pages/competition/index.ts
@@ -1,3 +1,1 @@
-export { CompetitionTabs } from "./CompetitionTabs";
-export { MyCompetitionPage } from "./my/MyCompetitionPage";
 export { RivalCompetitionPage } from "./rival/RivalCompetitionPage";
