--- apps/mac/src/renderer/src/pages/competition/my/MyCompetitionPage.style.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-`;
