--- apps/mac/src/renderer/src/pages/competition/my/MyCompetitionPage.tsx
@@ -1,10 +1,0 @@
-import * as S from "./MyCompetitionPage.style";
-import { WithMyCompetition } from "@/features/competition";
-
-export const MyCompetitionPage = () => {
-  return (
-    <S.Wrapper>
-      <WithMyCompetition />
-    </S.Wrapper>
-  );
-};
