+++ apps/mac/src/renderer/src/pages/competition/rival/RivalCompetitionPage.style.ts
@@ -9,6 +9,16 @@ export const Wrapper = styled.div`
   gap: 1rem;
 `;
 
+export const EmptyState = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+`;
+
 export const EmptyText = styled.p`
   ${font.headline2.medium}
   color: ${({ theme }) => theme.label.neutral};
